requirements.txt
python-dotenv
requests
psycopg2-binary
openai
numpy
